requirements.txt
anyjson==0.3.3
appdirs==1.4.3
arrow==0.12.1
astroid==1.6.3
beautifulsoup4==4.6.0
billiard==3.5.0.3
certifi==2018.4.16
chardet==3.0.4
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
Django-2.1.3
django-daterange-filter==1.3.0
django-extensions==2.1.0
django-tinymce==2.7.0
django-user-agents==0.3.2
djangorestframework==3.8.2
easy-thumbnails==2.5
flake8==3.5.0
future==0.16.0
gunicorn==19.9.0
idna==2.6
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
olefile==0.45.1
ordered-set==2.0.2
packaging==17.1
phonenumbers==8.9.4
Pillow==5.1.0
psycopg2==2.7.5
pycodestyle==2.4.0
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.4
pyparsing==2.2.0
pyTelegramBotAPI==3.6.3
python-dateutil==2.7.2
python-Levenshtein==0.12.0
pytz==2018.4
PyYAML==3.12
qrcode==6.0
raven==6.9.0
redis==2.10.6
requests==2.18.4
setproctitle==1.1.10
six==1.11.0
sorl-thumbnail==12.4.1
ua-parser==0.8.0
uritemplate==3.0.0
urllib3==1.22
user-agents==1.1.0
uuid==1.30
viberbot==1.0.11
vine==1.1.4
wallet-py3k==0.0.3
wrapt==1.10.11
yandex-checkout==1.0.3
PyPDF2==1.26.0
python-barcode==0.9.0
reportlab==3.5.4
numpy==1.15.1
pandas==0.23.4
et-xmlfile==1.0.1
jdcal==1.4
openpyxl==2.5.5
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
dawg-python==0.7.2
docopt==0.6.2
pyfcm==1.4.5
requests-toolbelt==0.8.0
yargy
pymorphy2==0.8
backports.functools-lru-cache==1.3
intervaltree==2.1.0
sortedcontainers
django-lazysignup-2.0.0